Added major changes due to errors!

+++ project1_starter.py
@@ -4,6 +4,7 @@ Name: [Chase Parker]
 Date: [31/10/25]
 
 AI Usage: [Document any AI assistance used]
+I used AI to solve the issues for load character and save character
 Example: AI helped with file I/O error handling logic in save_character function
 """
 import os
@@ -31,19 +32,21 @@ def calculate_stats(character_class, level):
         health_equation=100-(25+strength)
         magic= ((strength+health)*booster)/2
         return strength,int(magic),health_equation
-    elif character_class=="Rouge":
+    elif character_class=="Rogue":
         strength=(level*5)
         health=100
         health_equation=100-(50+strength)
         magic= (strength+(health*booster))/4
         return strength,int(magic),health_equation
-    elif character_class=="Clerics":
+    elif character_class=="Cleric":
         strength=(level*5)
         health=100
         health_equation=100-(25+strength)
         magic= (strength+health*booster)/2
         return strength,int(magic),health_equation
     
+    else:
+        return (0,0,0)
 
     # TODO: Implement this function
     # Return a tuple: (strength, magic, health)
@@ -59,13 +62,13 @@ def create_character(name, character_class):
     """
     strength,magic,health=calculate_stats(character_class,1)
     dictionary= {
-        "chr_name":name,
-        "chr_class":character_class,
-        "level":1,
-        "strength":strength,
-        "magic":magic,
-        "health":health,
-        "gold":100
+        'name':name,
+        'class':character_class,
+        'level':1,
+        'strength':strength,
+        'magic':magic,
+        'health':health,
+        'gold':100
     }
     return dictionary
     # TODO: Implement this function
@@ -86,19 +89,23 @@ def save_character(character, filename):
     Health: [health]
     Gold: [gold]
     """
+    if "/" in filename and not os.path.exists(filename):
+        return False
     with open(filename,'w') as file:
-       file.write(f"Character Name: {character['chr_name']}\n")
-       file.write(f"Class: {character['chr_class']}\n")
+        
+       file.write(f"Character Name: {character['name']}\n")
+       file.write(f"Class: {character['class']}\n")
        file.write(f"Level: {character['level']}\n")
        file.write(f"Strength: {character['strength']}\n")
        file.write(f"Magic: {character['magic']}\n")
        file.write(f"Health: {character['health']}\n")
        file.write(f"Gold: {character['gold']}")
-  
+    
+    
+    
     if os.path.isfile(filename):
        return True
-    else:
-       return False
+   
     
 
     # TODO: Implement this function
@@ -110,9 +117,30 @@ def load_character(filename):
     Loads character from text file
     Returns: character dictionary if successful, None if file not found
     """
+    if not os.path.isfile(filename):
+        return None
     with open(filename,'r') as file:
-       file_read=file.read()
-       return file_read
+       chr_dict={}
+       file_read=file.readlines()
+       for i in file_read:
+            t=i.split(":")
+            key=t[0]
+            val=t[1].strip()
+            if key=="Character Name":
+                chr_dict["name"]=val
+            elif key=="Class":
+                chr_dict["class"]=val
+            elif key=="Level":
+                chr_dict["level"]=int(val)
+            elif key=="Strength":
+                chr_dict["strength"]=int(val)
+            elif key=="Magic":
+                chr_dict["magic"]=int(val)
+            elif key=="Health":
+                chr_dict["health"]=int(val) 
+            elif key=="Gold":
+                chr_dict["gold"]=int(val)
+       return chr_dict
     # TODO: Implement this function
     # Remember to handle file not found errors
     
@@ -133,7 +161,7 @@ def display_character(character):
     Gold: 100
     """
     print("=== CHARACTER SHEET ===")
-    print((f"Class: {character['chr_class']}"))
+    print((f"Class: {character['class']}"))
     print(f"Level: {character['level']}")
     print(f"Strength: {character['strength']}")
     print(f"Magic: {character['magic']}")
@@ -149,14 +177,13 @@ def level_up(character):
     Modifies the character dictionary directly
     Returns: None
     """
-    level=1
+    
     if character['level']>=1:
-        level+=1
-        character['level']=level
-        character['strength']=character['strength']*level
-        character['magic']=character['magic']*level
-        character['health']=character['health']*level
-        character['gold']=character['gold']*level
+        character['level']+=1
+        character['strength']=character['strength']*character['level']
+        character['magic']=character['magic']*character['level']
+        character['health']=character['health']*character['level']
+        character['gold']=character['gold']*character['level']
         return character
     # TODO: Implement this function
     # Remember to recalculate stats for the new level
@@ -168,14 +195,15 @@ if __name__ == "__main__":
     print("Test your functions here!")
     
     # Example usage:
-    char = create_character("TestHero", "Warrior")
+    name_input=input()
+    class_input=input()
+    char = create_character(name_input, class_input)
     display_character(char)
-    print('hi')
-    level_up(char)
+    #level_up(char)
+    #display_character(char)
+    #level_up(char)
     display_character(char)
-    level_up(char)
-    display_character(char)
-    #save_character(char, "my_character.txt")
-    #loaded = load_character("my_character.txt")
+    save_character(char, "character.txt")
+    loaded = load_character("character.txt")
